Move clue widget menu actions to separate class
