+++ src/main/java/com/cluedetails/ClueInventoryManager.java
@@ -210,48 +210,6 @@ public class ClueInventoryManager
 		return cluesInInventory.get(clueItemID);
 	}
 
-	public void addHighlightWidgetSubmenus(MenuEntry[] entries, CluePreferenceManager cluePreferenceManager)
-	{
-		if (!client.isKeyPressed(KeyCode.KC_SHIFT))
-		{
-			return;
-		}
-
-		if (cluesInInventory.isEmpty())
-		{
-			return;
-		}
-
-		for (int idx = entries.length - 1; idx >= 0; --idx)
-		{
-			MenuEntry entry = entries[idx];
-
-			Widget widget = entry.getWidget();
-			if (widget == null || widget.getActions() == null)
-			{
-				return;
-			}
-
-			// Disable item marking, separate feature
-			if (widget.getItemId() != -1)
-			{
-				continue;
-			}
-
-			int componentId = widget.getId();
-			int childIndex = widget.getIndex();
-			WidgetId widgetId = new WidgetId(componentId, childIndex == -1 ? null : childIndex);
-
-			MenuEntry clueDetailsEntry = client.getMenu().createMenuEntry(1) // place above Cancel
-					.setOption("Clue details")
-					.setType(MenuAction.RUNELITE);
-			Menu submenu = clueDetailsEntry.createSubMenu();
-
-			cluesInInventory.forEach((id, instance) -> instance.getClueIds().forEach((clueId) -> addHighlightWidgetMenu(cluePreferenceManager, submenu, Clues.forClueIdFiltered(clueId), widgetId)));
-			break;
-		}
-	}
-
 	public void onMenuEntryAdded(MenuEntryAdded event, CluePreferenceManager cluePreferenceManager, ClueDetailsParentPanel panel)
 	{
 		MenuEntry entry = event.getMenuEntry();
@@ -422,23 +380,6 @@ public class ClueInventoryManager
 				updateClueItems(clue, itemId, cluePreferenceManager));
 	}
 
-	private void addHighlightWidgetMenu(CluePreferenceManager cluePreferenceManager, Menu menu, Clues clue, WidgetId widgetId)
-	{
-		if (clue == null) return;
-
-		boolean widgetInCluePreference = cluePreferenceManager.widgetsPreferenceContainsWidget(clue.getClueID(), widgetId);
-
-		String action = widgetInCluePreference ? "Remove from " : "Add to ";
-		String clueDetail = clue.getDetail(configManager);
-		final String text = action + "'" + clueDetail + "'";
-
-		// Add menu to widget for clue
-		menu.createMenuEntry(-1)
-				.setOption(text)
-				.setType(MenuAction.RUNELITE)
-				.onClick(e -> updateClueWidgets(clue, widgetId, cluePreferenceManager));
-	}
-
 	private void toggleMarkClue(CluePreferenceManager cluePreferenceManager, ClueDetailsParentPanel panel, int clueId, boolean isMarked, String target)
 	{
 		// We don't want to have marking on masters I think
@@ -479,32 +420,6 @@ public class ClueInventoryManager
 
 		// Save Clue itemIds
 		cluePreferenceManager.saveItemsPreference(clueId, clueItemIds);
-	}
-
-	private void updateClueWidgets(Clues clue, WidgetId widgetId, CluePreferenceManager cluePreferenceManager)
-	{
-		// Get existing Clue widgetIds
-		int clueId = clue.getClueID();
-		List<WidgetId> clueWidgetIds = cluePreferenceManager.getWidgetsPreference(clueId);
-
-		if (clueWidgetIds == null)
-		{
-			clueWidgetIds = new ArrayList<>();
-		}
-
-		// Remove if already present
-		if (clueWidgetIds.contains(widgetId))
-		{
-			clueWidgetIds.remove(widgetId);
-		}
-		// Add if not present
-		else
-		{
-			clueWidgetIds.add(widgetId);
-		}
-
-		// Save Clue widgetIds
-		cluePreferenceManager.saveWidgetsPreference(clueId, clueWidgetIds);
 	}
 
 	private boolean hasClueName(String name)
